monitor/requirements.txt
telethon>=1.24.0
pyyaml>=5.4.1
pytz>=2021.3